+++ actix-web-lab/src/middleware.rs
@@ -2,11 +2,12 @@
 //!
 //! Analogous to the `middleware` module in Actix Web.
 
+#[allow(deprecated)]
+pub use crate::middleware_from_fn::{from_fn, MiddlewareFn, Next};
 pub use crate::{
     catch_panic::CatchPanic,
     err_handler::ErrorHandlers,
     load_shed::LoadShed,
-    middleware_from_fn::{from_fn, MiddlewareFn, Next},
     middleware_map_response::{map_response, MapResMiddleware},
     middleware_map_response_body::{map_response_body, MapResBodyMiddleware},
     normalize_path::NormalizePath,