+++ actix-web-lab/src/redirect_to_www.rs
@@ -53,8 +53,10 @@ mod tests {
     };
 
     use super::*;
+    #[allow(deprecated)]
     use crate::middleware::from_fn;
 
+    #[allow(deprecated)]
     fn test_app() -> App<
         impl ServiceFactory<
             ServiceRequest,