+++ actix-web-lab/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Implement `Clone` for `extract::Path<T: Clone>`.
 - The `Deref` implementation for `header::CacheControl` now returns a slice instead of a `Vec`.
+- Deprecate `middleware::from_fn()` now it has graduated to Actix Web.
 - Deprecate `extract::ThinData` now it has graduated to Actix Web.
 
 ## 0.21.0